requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-core
langchain-openai
openai
python-multipart
python-dotenv
langserve
tiktoken
-e .